Create README.md
This is the First version of a Windows c# program that calculates a solar PV system with battery. Several loads from heat pump to industry profiles are included. Peak power and battery can be opimized for lowes cost (LCOE) You need Scottplot Nuget Pakage. This allows you to Zoom all Graphs interactive.

PVCalc © 2025 by Prof. Dr.-Ing. Thomas Hücker from HTW Berlin

This program is free software: you can redistribute it and/or modify it under the terms of the GNU General Public License as published by the Free Software Foundation, either version 3 of the License, or (at your option) any later version.

This program is distributed in the hope that it will be useful,but WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.See theGNU General Public License for more details. You should have received a copy of the GNU General Public License along with this program. If not, see: http://www.gnu.org/licenses/